Rest-API-JWT/requirements.txt
flask
psycopg2-binary
python-dotenv
PyJWT